Adds probes, but for Kafka I don't think it indicates readiness...
which might not matter because we no longer have a loadbalancing service.

These probes won't catch all failure modes,
but if they fail we're pretty sure the container is malfunctioning.

I found some sources recommending ./bin/kafka-topics.sh for probes
but to me it looks risky to introduce a dependency to some other service for such things.
One such source is https://github.com/kubernetes/charts/pull/144

The zookeeper probe is from
https://kubernetes.io/docs/tasks/configure-pod-container/configure-liveness-readiness-probes/
An issue is that zookeeper's logs are quite verbose for every probe.