Removes some "platform" ideas from readme as they need real PRs
rather than a couple of sentences. In response to #103.

Regarding Kafka Streams I no longer think we need examples,
because any dockerized streams application can run as a deployment.

An example of Kafka Connect would be useful, in particular
the combination of a custom image (or one with stock connectors)
and a Connect cluster manifest.
This is tracked in #69, but not the relation to
https://github.com/solsson/dockerfiles/tree/master/connect-*

KSQL (#68) is highly interesting.