Auto create topics, in our experience, creates more pains than it solves
so we want to undo https://github.com/Yolean/kubernetes-kafka/pull/107.
It was partially based on a false assumption, as pointed out in
https://github.com/Yolean/kubernetes-kafka/issues/101#issuecomment-365600781

Topics are created not only at produce but also at for example kafkacat -C.
Typos cost us more time than it would take to automate topic creation
and run ./bin/kafka-topics.sh in a temporary pod when we haven't automated.