Topics now get 12 partitions by default
The number 12 was chosen to support even load.
Counts of 2, 3, 4 and 6 are reasonable for both replicas and consumer group members.
At the same time we didn't want a high default number because the sample config file
mentions "more files across the brokers".

The reason for doing this change is that clients should not assume ordering guarantees unless designed for.