There's no reason to ever recommend auto.create.topics=true
I though that the feature would be neat for development (with replication factor 1, see #222)
but it causes just as much confusion and useless troubleshooting there,
for example race conditions between intentional topic creation and a container
starting up to produce to the topic. You actually never know which topic config you're getting.

Related: #107

The duplication is a workaround for https://github.com/kubernetes-sigs/kustomize/issues/642