Breaking: renames the statefulset's headless service so
the pod name or statefulset.kubernetes.io/pod-name can be used via
the downward api in args to do things like
--override listeners=PLAINTEXT://$(POD_NAME).kafka:9092

Once again it's unfortunate that the statefulset label is pod name,
not pod index.

Also makes sure that DNS entries are published prior to readiness
so clusters don't get into loops of not being able to find each other.